Initial commit for the Shopping list generator
This application is a node.js Shopping list generator. It outputs in
two formats. The first is a plain web page so the user can see the
list. The second is a plain text file in the lists/ directory. This
file is in simple Emacs Org-mode format for a check list. This allows
me to send it to my phone and use it as a checklist while shopping.
The Readme.md file will have the details of the files needed and
where everything goes.

Changes to be committed:
	new file:   app.js
	new file:   backups/
	new file:   inventory.json
	new file:   lists/
	new file:   package.json
	new file:   public/script.js
	new file:   public/styles.css
	new file:   public/sytles.css
	new file:   ssl/
	new file:   tests/app.test.js
	new file:   views/index.ejs
	new file:   views/list.ejs