add bootstrap css support and rewrite generate-list
Asked ChatGPT4 to improve the UI and it respondede with rewriting
app.js to support bootstrap css. But that failed in the
browser (behind the filewall) so I pulled back a copy and removed  the
SHA signature. Also had to rewrite generate-list as it didn't include
that code in the redesign (UI Improvement). While the web interface
does look a little better, it was a lot of breakage. I think ChatGPT4
is starting to hallucinate. Other things broke also.