Update to switch to bootstrap css and add a delete item buttom
This is a total rewite of the main index.ejs page and code. It was
meant to improve the UI but seems unnecessary to this extent. I went
with it and tried to take advantage of the changes. It did require
changes to the generate-list code.  I also asked for a new feature, a
delete key. I asked in a vauge manner and it gave me a delete on each
item. I also had to add back in the aisle color code. I chose not to
ask ChatGPT4 for this code.